Revert "fixed to show recipe name"
This reverts commit d0a62bdca62121c0dce1157bf2580814b2ee4da1.

+++ Recipe Book/RecipeBook/Migrations/RecipeBookContextModelSnapshot.cs
@@ -30,8 +30,8 @@ namespace RecipeBook.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<double>("Quantity")
-                        .HasColumnType("float");
+                    b.Property<int>("Quantity")
+                        .HasColumnType("int");
 
                     b.Property<int>("RecipeId")
                         .HasColumnType("int");
